Let's see if I can still C#
Initial commit, with basic C# program. I want to make a program that
downloads all my plays from Last.fm, cross-references them with my music
library, and then computes statistics such as most popular track. The
goal is to find specific subsets of my library, so I can build playlists
that fit on a device with limited storage capacity (say, 64 GiB).

I need a language with good http and xml or json support for this.
Haskell would be a good fit, but they recently broke everything (again)
and now there is no GHC for my system that Stack can set up. Python
would be fine but types are nice. And C# is actually quite a nice
language. (Or maybe I have just forgotton all the pain points over time
-- let's see how I feel about this when the first NullReferenceException
occurs.)

So I spent 15 minutes trying to get 'dotnet' working on my machine,
because it is supposed to be cross-platform now. No success. Let's use
Mono then, it just works.

Maybe I'll regret this, there are so many monoids around that just beg
to be abstracted over: getting the first play for every song and the
last play is exactly the same, only in the first case we combine with
min, and in the second case we combine with max. And then if you have
those play counts per track, you can get them per album and per artist
by combining them with the monoid operation. Total play count is another
instance that works in *exactly* the same way. Hmmm. Ok, maybe I should
try a bit harder to get a working GHC again. Or maybe I should not whine
and try to get a program running before thinking about monoids.