Compute which months to retrieve
This Last.fm API is a bit unfortunate, because it paginates by reverse
chronological ordering. This means that page numbers are not stable if
you do not provide an upper bound on the time. So page numbers are not
something we can use to identify a time span. What does work is the time
span itself. There would be multiple pages per time span, but we can
retrieve those once. This would only need to be re-done for the last
time span, all the ones in the past are guaranteed not to change.

So how big should a time span be? There are a few considerations:

 * Retrieval of multiple spans can be parallelized, because the bounds
   are known. But the pages in the span have to be retrieved
   sequentially. Or at least, one page has to be retrieved before we
   know the total number of pages, after which we can do things in
   parallel. So more spans means better parallelism.

 * Short spans are undesirable because they would cause many files to be
   created on the file system, if every file is stored in one span.

 * Short spans are undesirable, because the limit of 200 scrobbles per
   page will not be reached, which means more calls than necessary have
   to be made.

 * Long spans are undesirable because the last span sometimes needs to
   be updated, and a long span takes longer to retrieve.

My current guess is that spans of one months would be good. I have
roughly 12000 scrobbles per year, so with 200 per page that is five
pages per month. Coming in at about a second per request, the time to
update the current month would be reasonable (after the initial download
of all history).